Added support for multiple answers, so that if a correct answer is given it also shows the other possibilities.
Also added in support for multiple choice questions, so that if a player guesses one of the answers they get locked out from answering again.  Players can still choose to skip a question at this point.

Also refactored the check_answer() method.  Its default behaviour is to award a point if it falls through to the end.  The if statements within it guard against all the conditions that would prevent a point from being awarded.

Closes #4.